# backend/docs/swagger.yaml
basePath: /
definitions:
...
  ai-models-backend_internal_models.TodoCreateRequest:
    properties:
      description:
        description: 可选
        type: string
      dueDate:
        description: 可选
        type: string
      position:
        description: 可选，不传则自动分配
        type: number
      priority:
        description: 可选
        type: integer
      title:
        description: 必填
        type: string
    required:
    - title
    type: object
  ai-models-backend_internal_models.TodoPositionItem:
    properties:
      id:
        description: 必填
        type: string
      position:
        description: 必填
        type: number
    required:
    - id
    - position
    type: object
  ai-models-backend_internal_models.TodoPositionUpdateRequest:
    properties:
      items:
        description: 必填，dive验证数组元素
        items:
          $ref: '#/definitions/ai-models-backend_internal_models.TodoPositionItem'
        type: array
    required:
    - items
    type: object
  ai-models-backend_internal_models.TodoResponse:
    properties:
      completed:
        type: boolean
      createdAt:
        type: string
      description:
        type: string
      dueDate:
        type: string
      id:
        type: string
      position:
        type: number
      priority:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
  ai-models-backend_internal_models.TodoUpdateRequest:
    properties:
      completed:
        description: 可选，使用指针区分false和未设置
        type: boolean
      description:
        description: 可选，不传不更新
        type: string
      dueDate:
        description: 可选，不传不更新
        type: string
      position:
        description: 可选，支持更新排序位置
        type: number
      priority:
        description: 可选，不传不更新
        type: integer
      title:
        description: 可选，不传不更新
        type: string
    type: object
...
  ai-models-backend_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: AI Models Backend API documentation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: AI Models API
  version: "1.0"
paths:
...
  /todos:
    get:
      description: 分页获取TODO列表，按position升序排列，支持按完成状态筛选
      operationId: getTodoList
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 完成状态
        in: query
        name: completed
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: 获取TODO列表
      tags:
      - TODO
    post:
      description: 创建新的TODO项，支持设置标题、描述、优先级、位置和截止日期。不指定position时自动分配到最后
      operationId: createTodo
      parameters:
      - description: 创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ai-models-backend_internal_models.TodoCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ai-models-backend_internal_models.TodoResponse'
              type: object
      summary: 创建TODO
      tags:
      - TODO
  /todos/{id}:
    delete:
      description: 根据ID永久删除指定的TODO项，操作不可逆
      operationId: deleteTodo
      parameters:
      - description: TODO ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: 删除TODO
      tags:
      - TODO
    get:
      description: 根据唯一标识符获取指定的TODO项详细信息
      operationId: getTodoByID
      parameters:
      - description: TODO ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ai-models-backend_internal_models.TodoResponse'
              type: object
      summary: 根据ID获取TODO
      tags:
      - TODO
    put:
      description: 根据ID更新现有TODO项的内容，支持更新标题、描述、完成状态、优先级、位置等
      operationId: updateTodo
      parameters:
      - description: TODO ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ai-models-backend_internal_models.TodoUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ai-models-backend_internal_models.TodoResponse'
              type: object
      summary: 更新TODO
      tags:
      - TODO
  /todos/{id}/toggle:
    patch:
      description: 快速切换TODO项的完成状态（完成/未完成）
      operationId: toggleTodoComplete
      parameters:
      - description: TODO ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ai-models-backend_internal_models.TodoResponse'
              type: object
      summary: 切换TODO完成状态
      tags:
      - TODO
  /todos/positions:
    put:
      description: 批量更新TODO项的位置，支持拖拽排序功能
      operationId: updateTodoPositions
      parameters:
      - description: 位置更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ai-models-backend_internal_models.TodoPositionUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: 批量更新TODO位置
      tags:
      - TODO
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"